cleaned up the SearchEventViewModel, SearchEventState in accordance to clean architecture and Hugo reviews
